added mypy support, mypy fixes, black-format files [CE-155]

+++ src/orcidlink/service_clients/auth.py
@@ -5,11 +5,12 @@ from orcidlink.service_clients import authclient2
 ##
 #
 
+
 def get_username(kbase_auth_token: str) -> str:
     auth = authclient2.KBaseAuth(
         auth_url=config().kbase.services.Auth2.url,
-        cache_lifetime=config().kbase.services.Auth2.tokenCacheLifetime / 1000,
-        cache_max_size=config().kbase.services.Auth2.tokenCacheMaxSize
+        cache_lifetime=int(config().kbase.services.Auth2.tokenCacheLifetime / 1000),
+        cache_max_size=config().kbase.services.Auth2.tokenCacheMaxSize,
     )
 
     return auth.get_username(kbase_auth_token)
